Epsilon dominance, domination factor, (unused) crowding (#63)
* feat: initial epsilon-pareto dominance

* feat: domination factor

Introduced domination factor as a mechanism to differentiate solutions that do not dominate each other. Crowding factor was also implemented, but is not currently used (does not improve the results)

* feat: epsilon dominance

Implemented epsilon-dominance in such a way that promotes the agent calling the function. If it dominates the other solution, it does not have the chance to check if it dominates the first one as well